add Batch Minting cleanup MintInfo
