+++ DotNut/ApiModels/BatchMint/PostBatchedMintRequest.cs
@@ -6,14 +6,14 @@ public class PostBatchedMintRequest
 {
     [JsonPropertyName("quotes")]
     public string[] QuoteIds { get; set; }
-    
+
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
-    [JsonPropertyName("amounts")]
+    [JsonPropertyName("quote_amounts")]
     public ulong?[]? Amounts { get; set; }
-    
+
     [JsonPropertyName("outputs")]
     public BlindedMessage[] Outputs { get; set; }
-    
+
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingDefault)]
     [JsonPropertyName("signatures")]
     public string?[]? Signatures { get; set; }
