+++ DotNut/ApiModels/BatchMint/PostBatchedMintResponse.cs
